Resolves #56 by moving kmk.common.* up a level to kmk.*
