+++ kmk/util.py
