+++ user_keymaps/klardotsh/feather_m4_express/fourfour.py
@@ -1,10 +1,10 @@
-from kmk.common.consts import DiodeOrientation, UnicodeModes
-from kmk.common.keycodes import KC
-from kmk.common.macros.simple import send_string, simple_key_sequence
-from kmk.common.macros.unicode import unicode_codepoint_sequence
-from kmk.common.pins import Pin as P
+from kmk.consts import DiodeOrientation, UnicodeModes
 from kmk.entrypoints.handwire.feather_m4_express import main
 from kmk.firmware import Firmware
+from kmk.keycodes import KC
+from kmk.macros.simple import send_string, simple_key_sequence
+from kmk.macros.unicode import unicode_codepoint_sequence
+from kmk.pins import Pin as P
 
 cols = (P.D11, P.D10, P.D9)
 rows = (P.A2, P.A3, P.A4, P.A5)
